Big Refactor, new hybrid way to generate the table and added configs
Refactored the most of the code to improve the base of the project and
legibility, now this repo is only for the library (temporary) and the
project will have 3 bigs repos, this one will have all the necessary
functions to create and draw Nucleide charts.
In this version the Nucleide Chart is formed by Element boxes that are a
group of one or more nucleids represented by their mass number and Decay
mode.
New Features:
   All the structure has been refactored to look more like the Karlsruhe
Nucleide Chart, the structure should be similar so a person able to read
the Karlsruhe Chart should be able to understand this one
   Now the csv driver can read and storage different nucleides with
similar characteristics, representing it on the final nucleide chart as
different sections of a same element box
   Config on config.py, this shouldn't be changed and is only for have a
base structure for the other configs
TODO:
   Create JSON configs equivalent to config.py
   Create a JSON config loader
   Make Nucleid Chart divisible (more likely to be slicer by rows, but
maybe i will set some configs to customize it)
   Make different divisible configs based on different standars and
orientations
   Make legend and other elements of the nucleide chart that doesn't
represents directly the nucleides